Separate rest of scraper to scraper.js
